fixed issues with adding an aside sidebar
